+++ x/callback/keeper/callback_test.go
@@ -2,6 +2,7 @@ package keeper_test
 
 import (
 	"fmt"
+	"strings"
 
 	sdk "github.com/cosmos/cosmos-sdk/types"
 
@@ -19,12 +20,19 @@ func (s *KeeperTestSuite) TestSaveCallback() {
 	keeper.SetWasmKeeper(contractViewer)
 	validCoin := sdk.NewInt64Coin("stake", 10)
 
-	contractAddr := e2eTesting.GenContractAddresses(1)[0]
+	contractAddresses := e2eTesting.GenContractAddresses(3)
+	contractAddr := contractAddresses[0]
+	contractAddr2 := contractAddresses[1]
+	contractAddr3 := contractAddresses[2]
 	contractAdminAcc := s.chain.GetAccount(0)
 	notContractAdminAcc := s.chain.GetAccount(1)
 	contractOwnerAcc := s.chain.GetAccount(2)
 	contractViewer.AddContractAdmin(
 		contractAddr.String(),
+		contractAdminAcc.Address.String(),
+	)
+	contractViewer.AddContractAdmin(
+		contractAddr2.String(),
 		contractAdminAcc.Address.String(),
 	)
 
@@ -35,6 +43,16 @@ func (s *KeeperTestSuite) TestSaveCallback() {
 	metaCurrent.RewardsAddress = contractOwnerAcc.Address.String()
 	rewardsKeeper.SetContractInfoViewer(contractViewer)
 	err := rewardsKeeper.SetContractMetadata(ctx, contractAdminAcc.Address, contractAddr, metaCurrent)
+	s.Require().NoError(err)
+
+	// Setting callback module as contract owner
+	blockedModuleAddr := s.chain.GetApp().Keepers.AccountKeeper.GetModuleAccount(ctx, types.ModuleName).GetAddress()
+	s.Require().True(s.chain.GetApp().Keepers.BankKeeper.BlockedAddr(blockedModuleAddr))
+	var metaCurrent2 rewardsTypes.ContractMetadata
+	metaCurrent2.ContractAddress = contractAddr2.String()
+	metaCurrent2.OwnerAddress = blockedModuleAddr.String()
+	metaCurrent2.RewardsAddress = contractAddr2.String()
+	err = rewardsKeeper.SetContractMetadata(ctx, contractAdminAcc.Address, contractAddr2, metaCurrent2)
 	s.Require().NoError(err)
 
 	params, err := keeper.GetParams(ctx)
@@ -66,7 +84,7 @@ func (s *KeeperTestSuite) TestSaveCallback() {
 		{
 			testCase: "FAIL: contract does not exist",
 			callback: types.Callback{
-				ContractAddress: e2eTesting.GenContractAddresses(2)[1].String(),
+				ContractAddress: contractAddr3.String(),
 				JobId:           1,
 				CallbackHeight:  101,
 				ReservedBy:      contractAddr.String(),
@@ -149,6 +167,23 @@ func (s *KeeperTestSuite) TestSaveCallback() {
 			errorType:   types.ErrCallbackHeightTooFarInFuture,
 		},
 		{
+			testCase: "FAIL: sender is a blocked address",
+			callback: types.Callback{
+				ContractAddress: contractAddr2.String(),
+				JobId:           1,
+				CallbackHeight:  102,
+				ReservedBy:      blockedModuleAddr.String(),
+				FeeSplit: &types.CallbackFeesFeeSplit{
+					TransactionFees:       &validCoin,
+					BlockReservationFees:  &validCoin,
+					FutureReservationFees: &validCoin,
+					SurplusFees:           &validCoin,
+				},
+			},
+			expectError: true,
+			errorType:   types.ErrUnauthorized,
+		},
+		{
 			testCase: "OK: save callback - sender is contract",
 			callback: types.Callback{
 				ContractAddress: contractAddr.String(),
@@ -187,6 +222,22 @@ func (s *KeeperTestSuite) TestSaveCallback() {
 				JobId:           3,
 				CallbackHeight:  101,
 				ReservedBy:      contractAdminAcc.Address.String(),
+				FeeSplit: &types.CallbackFeesFeeSplit{
+					TransactionFees:       &validCoin,
+					BlockReservationFees:  &validCoin,
+					FutureReservationFees: &validCoin,
+					SurplusFees:           &validCoin,
+				},
+			},
+			expectError: false,
+		},
+		{
+			testCase: "OK: save callback - sender is contract admin but address is in uppercase",
+			callback: types.Callback{
+				ContractAddress: contractAddr.String(),
+				JobId:           1,
+				CallbackHeight:  102,
+				ReservedBy:      strings.ToUpper(contractAdminAcc.Address.String()),
 				FeeSplit: &types.CallbackFeesFeeSplit{
 					TransactionFees:       &validCoin,
 					BlockReservationFees:  &validCoin,
@@ -331,8 +382,7 @@ func (s *KeeperTestSuite) TestDeleteCallback() {
 				CallbackHeight:  101,
 				ReservedBy:      contractAddr.String(),
 			},
-			expectError: true,
-			errorType:   types.ErrCallbackNotFound,
+			expectError: false, // Should silently fail. MsgSrvr ensures that callback exists before calling keeper
 		},
 		{
 			testCase: "OK: Success delete - sender is contract",
@@ -367,7 +417,7 @@ func (s *KeeperTestSuite) TestDeleteCallback() {
 	}
 	for _, tc := range testCases {
 		s.Run(fmt.Sprintf("Case: %s", tc.testCase), func() {
-			err := keeper.DeleteCallback(ctx, tc.callback.ReservedBy, tc.callback.CallbackHeight, tc.callback.ContractAddress, tc.callback.JobId)
+			err := keeper.DeleteCallback(ctx, tc.callback.ReservedBy, tc.callback)
 			if tc.expectError {
 				s.Require().Error(err)
 				s.Assert().ErrorContains(err, tc.errorType.Error())