+++ CHANGELOG.md
@@ -54,6 +54,7 @@ Contains all the PRs that improved the code without changing the behaviors.
 ### Fixed
 - [#537](https://github.com/archway-network/archway/pull/537) - Fix issue with callback failing when module param is changed
 - [#538](https://github.com/archway-network/archway/pull/538) - Fixing the interchain test gh workflow failing cuz rc tags were not recognized 
+- [#539](https://github.com/archway-network/archway/pull/539) - Remediations for x/callback audit
 
 
 ## [v6.0.0](https://github.com/archway-network/archway/releases/tag/v6.0.0)